Simplified the code for the renderer mixin
This commit adds a parameter to the wolf variants enum that specifies the name of the wolf variant, which the renderer uses to get the texture more easily.
I've also removed the pale wolf's textures as they didn't seem necessary to include, but if that is a problem they can simply be added back.